+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - support for Kotlin
+- preconditions which observe mdib changes during the test run
 
 ### Changed